+++ lib/SIL/SILValue.cpp
@@ -34,168 +34,6 @@ static_assert(sizeof(SILValue) == sizeof(uintptr_t),
 //                              Utility Methods
 //===----------------------------------------------------------------------===//
 
-static bool isRCIdentityPreservingCast(ValueKind Kind) {
-  switch (Kind) {
-  case ValueKind::UpcastInst:
-  case ValueKind::UncheckedRefCastInst:
-  case ValueKind::UncheckedRefCastAddrInst:
-  case ValueKind::UnconditionalCheckedCastInst:
-  case ValueKind::RefToBridgeObjectInst:
-  case ValueKind::BridgeObjectToRefInst:
-    return true;
-  default:
-    return false;
-  }
-}
-
-/// Return the underlying SILValue after stripping off identity SILArguments if
-/// we belong to a BB with one predecessor.
-static SILValue stripSinglePredecessorArgs(SILValue V) {
-  while (true) {
-    auto *A = dyn_cast<SILArgument>(V);
-    if (!A)
-      return V;
-
-    SILBasicBlock *BB = A->getParent();
-
-    // First try and grab the single predecessor of our parent BB. If we don't
-    // have one, bail.
-    SILBasicBlock *Pred = BB->getSinglePredecessor();
-    if (!Pred)
-      return V;
-
-    // Then grab the terminator of Pred...
-    TermInst *PredTI = Pred->getTerminator();
-
-    // And attempt to find our matching argument.
-    //
-    // *NOTE* We can only strip things here if we know that there is no semantic
-    // change in terms of upcasts/downcasts/enum extraction since this is used
-    // by other routines here. This means that we can only look through
-    // cond_br/br.
-    //
-    // For instance, routines that use stripUpcasts() do not want to strip off a
-    // downcast that results from checked_cast_br.
-    if (auto *BI = dyn_cast<BranchInst>(PredTI)) {
-      V = BI->getArg(A->getIndex());
-      continue;
-    }
-
-    if (auto *CBI = dyn_cast<CondBranchInst>(PredTI)) {
-      if (SILValue Arg = CBI->getArgForDestBB(BB, A)) {
-        V = Arg;
-        continue;
-      }
-    }
-
-    return V;
-  }
-}
-
-SILValue SILValue::stripCasts() {
-  SILValue V = *this;
-
-  while (true) {
-    V = stripSinglePredecessorArgs(V);
-
-    auto K = V->getKind();
-    if (isRCIdentityPreservingCast(K)
-        || K == ValueKind::UncheckedTrivialBitCastInst
-        || K == ValueKind::MarkDependenceInst) {
-      V = cast<SILInstruction>(V.getDef())->getOperand(0);
-      continue;
-    }
-
-    return V;
-  }
-}
-
-SILValue SILValue::stripUpCasts() {
-  assert(getType().isClassOrClassMetatype() &&
-         "Expected class or class metatype!");
-
-  SILValue V = stripSinglePredecessorArgs(*this);
-
-  while (isa<UpcastInst>(V))
-    V = stripSinglePredecessorArgs(cast<UpcastInst>(V)->getOperand());
-
-  return V;
-}
-
-SILValue SILValue::stripClassCasts() {
-  SILValue V = *this;
-  while (true) {
-    if (auto *UI = dyn_cast<UpcastInst>(V)) {
-      V = UI->getOperand();
-      continue;
-    }
-
-    if (auto *UCCI = dyn_cast<UnconditionalCheckedCastInst>(V)) {
-      V = UCCI->getOperand();
-      continue;
-    }
-
-    return V;
-  }
-}
-
-
-SILValue SILValue::stripAddressProjections() {
-  SILValue V = *this;
-
-  while (true) {
-    V = stripSinglePredecessorArgs(V);
-
-    switch (V->getKind()) {
-    case ValueKind::StructElementAddrInst:
-    case ValueKind::TupleElementAddrInst:
-    case ValueKind::RefElementAddrInst:
-    case ValueKind::ProjectBoxInst:
-    case ValueKind::UncheckedTakeEnumDataAddrInst:
-      V = cast<SILInstruction>(V.getDef())->getOperand(0);
-      continue;
-    default:
-      return V;
-    }
-  }
-}
-
-SILValue SILValue::stripValueProjections() {
-  SILValue V = *this;
-
-  while (true) {
-    V = stripSinglePredecessorArgs(V);
-
-    switch (V->getKind()) {
-    case ValueKind::StructExtractInst:
-    case ValueKind::TupleExtractInst:
-      V = cast<SILInstruction>(V.getDef())->getOperand(0);
-      continue;
-    default:
-      return V;
-    }
-  }
-}
-
-SILValue SILValue::stripIndexingInsts() {
-  SILValue V = *this;
-  while (true) {
-    if (!isa<IndexingInst>(V.getDef()))
-      return V;
-    V = cast<IndexingInst>(V)->getBase();
-  }
-}
-
-SILValue SILValue::stripExpectIntrinsic() {
-  SILValue V = *this;
-  auto *BI = dyn_cast<BuiltinInst>(V);
-  if (!BI)
-    return V;
-  if (BI->getIntrinsicInfo().ID != llvm::Intrinsic::expect)
-    return V;
-  return BI->getArguments()[0];
-}
-
 SILBasicBlock *ValueBase::getParentBB() {
   if (auto Inst = dyn_cast<SILInstruction>(this))
     return Inst->getParent();
@@ -203,53 +41,3 @@ SILBasicBlock *ValueBase::getParentBB() {
     return Arg->getParent();
   return nullptr;
 }
-
-void Operand::hoistAddressProjections(SILInstruction *InsertBefore,
-                                      DominanceInfo *DomTree) {
-  SILValue V = get();
-  SILInstruction *Prev = nullptr;
-  auto *InsertPt = InsertBefore;
-  while (true) {
-    SILValue Incoming = stripSinglePredecessorArgs(V);
-
-    // Forward the incoming arg from a single predecessor.
-    if (V != Incoming) {
-      if (V == get()) {
-        // If we are the operand itself set the operand to the incoming
-        // argument.
-        set(Incoming);
-        V = Incoming;
-      } else {
-        // Otherwise, set the previous projections operand to the incoming
-        // argument.
-        assert(Prev && "Must have seen a projection");
-        Prev->setOperand(0, Incoming);
-        V = Incoming;
-      }
-    }
-
-    switch (V->getKind()) {
-    case ValueKind::StructElementAddrInst:
-    case ValueKind::TupleElementAddrInst:
-    case ValueKind::RefElementAddrInst:
-    case ValueKind::UncheckedTakeEnumDataAddrInst: {
-      auto *Inst = cast<SILInstruction>(V);
-      // We are done once the current projection dominates the insert point.
-      if (DomTree->dominates(Inst->getParent(), InsertBefore->getParent()))
-        return;
-
-      // Move the current projection and memorize it for the next iteration.
-      Prev = Inst;
-      Inst->moveBefore(InsertPt);
-      InsertPt = Inst;
-      V = Inst->getOperand(0);
-      continue;
-    }
-    default:
-      assert(DomTree->dominates(V->getParentBB(), InsertBefore->getParent()) &&
-             "The projected value must dominate the insertion point");
-      return;
-    }
-  }
-}
-
