+++ lib/SIL/SILValueProjection.cpp
@@ -12,6 +12,7 @@
 
 #define DEBUG_TYPE "sil-value-projection"
 #include "swift/SIL/SILValueProjection.h"
+#include "swift/SIL/InstructionUtils.h"
 #include "llvm/Support/Debug.h"
 
 using namespace swift;
