+++ lib/SILOptimizer/Utils/CheckedCastBrJumpThreading.cpp
@@ -2,6 +2,7 @@
 #include "llvm/ADT/SmallPtrSet.h"
 #include "llvm/ADT/SmallVector.h"
 #include "swift/SIL/SILInstruction.h"
+#include "swift/SIL/InstructionUtils.h"
 #include "swift/SILOptimizer/Analysis/DominanceAnalysis.h"
 #include "swift/SILOptimizer/Utils/CFG.h"
 #include "swift/SILOptimizer/Utils/Local.h"
@@ -213,10 +214,10 @@ SILValue CheckedCastBrJumpThreading::isArgValueEquivalentToCondition(
     SILValue Value, SILBasicBlock *DomBB, SILValue DomValue,
     DominanceInfo *DT) {
   SmallPtrSet<ValueBase *, 16> SeenValues;
-  DomValue = DomValue.stripClassCasts();
+  DomValue = stripClassCasts(DomValue);
 
   while (true) {
-    Value = Value.stripClassCasts();
+    Value = stripClassCasts(Value);
     if (Value == DomValue)
       return Value;
 
@@ -241,7 +242,7 @@ SILValue CheckedCastBrJumpThreading::isArgValueEquivalentToCondition(
       // Each incoming value should be either from a block
       // dominated by DomBB or it should be the value used in
       // condition in DomBB
-      Value = IncomingValue.stripClassCasts();
+      Value = stripClassCasts(IncomingValue);
       if (Value == DomValue)
         continue;
 
@@ -578,7 +579,7 @@ bool CheckedCastBrJumpThreading::trySimplify(TermInst *Term) {
   // Init information about the checked_cast_br we try to
   // jump-thread.
   BB = Term->getParent();
-  Condition = Term->getOperand(0).stripClassCasts();
+  Condition = stripClassCasts(Term->getOperand(0));
   SuccessBB = CCBI->getSuccessBB();
   FailureBB = CCBI->getFailureBB();
 
@@ -636,7 +637,7 @@ bool CheckedCastBrJumpThreading::trySimplify(TermInst *Term) {
     // based on the found dominating checked_cast_br.
     DomSuccessBB = DomCCBI->getSuccessBB();
     DomFailureBB = DomCCBI->getFailureBB();
-    DomCondition = DomTerm->getOperand(0).stripClassCasts();
+    DomCondition = stripClassCasts(DomTerm->getOperand(0));
 
     // Init state variables for paths analysis
     SuccessPreds.clear();
